+++ invoicer_project/clients/templates/clients/client_add.html
@@ -100,6 +100,6 @@
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.1/jquery.min.js"></script>
 <script src="{% static 'js/index.js' %}"></script>
 <script src="{% static 'js/bundle.js' %}"></script>
-<script src="{% static 'js/add_client_request.js' %}"></script>
+<script type="module" src="{% static 'js/add_client_request.js' %}"></script>
 </body>
 </html>
