+++ invoicer_project/clients/templates/clients/client_edit.html
@@ -98,6 +98,6 @@
 
         <script src="{% static 'js/index.js' %}"></script>
         <script src="{% static 'js/bundle.js' %}"></script>
-        <script src="{% static 'js/edit_client_request.js' %}"></script>
+        <script type="module" src="{% static 'js/edit_client_request.js' %}"></script>
     </body>
 </html>
